Main change is addition of Integrator file, and restructing to accomodate the new Integrator class. Have not yet implemented it. Some minor refactoring done as well.

+++ install/include/nbody/constants.h
@@ -6,7 +6,7 @@
 namespace nbody {
 
   extern const size_t MAX_BODIES_RECOMMENDED; // = 10000
-  extern const float NEWTON_G; // = 1.0f
+  extern const double NEWTON_G; // = 1.0
 
 } // namespace nbody
 #endif // _NBODY_CONSTANTS_H
