+++ src/constants.cpp
@@ -2,5 +2,5 @@
 
 namespace nbody {
   const size_t MAX_BODIES_RECOMMENDED = 10000;
-  const float NEWTON_G = 1.0f;
+  const double NEWTON_G = 1.0;
 } // namespace nbody
